move settings components into settings folder
